Infection Data for Sigma Virus in Drosophila melanogaster

Infection data for sigma virus isolates across lines of Drosophila melanogaster

Usage

data(sigma)

Format

a data frame with 1092 rows and 4 columns. infected and not_infected give the number of flies in a vial that are infected or not infected with sigma virus. The viruses belong to one of five strains designated by their country of origin: United States (AP30), Spain (GC20), Greece (PF115), the United Kingdom (E27) and France (Hap23). The flies are genetically identical except at the third chromosome for which they belong to one of 67 lines line.

References

Carpenter, J.A et al. 2012 Evolution 66 1114–1125

See also