Diet composition of Tufted Puffin

Diet composition of Tufted Puffin (Fratercula cirrhata) chicks from Middleton Island in the Gulf of Alaska (1978-2024).

Usage

data(tufted_puffin)

Format

a data frame with 565 rows and 7 columns. Each row corresponds to a sample and year and month refer to the date of collection. The following four columns are the number of Capelin (Capelin), Pacific Herring (Herring), Walleye pollock (Pollock) and Pacific Sand lance (SandLance) in each sample. An additinoal 40 prey taxa were also recorded and their combined counts per sample are in the final column (Other).

Source

Hatch, S., Arimitsu, M., Piatt, J., Whelan, S., and Marsteller, C.E., 2023, Seabird Diet Data Collected on Middleton Island, Gulf of Alaska (ver 3.0, February 2025): U.S. Geological Survey data release, https://doi.org/10.5066/P93I0P67.